Adding some more code to Bricks - profile for interrupts - profile for syscalls Think about how to configure these profiles and try to register handlers

+++ crates/bricks/src/syscalls.rs
@@ -1,7 +1,7 @@
-/// Our own syscalls 
-pub const ATTEST_ENCLAVE : u32 = 1;
+/// Our own syscalls
+pub const ATTEST_ENCLAVE: usize = 1;
 
-/// Linux syscalls 
-pub const LINUX_MALLOC : u32 = 2;
+/// Linux syscalls
+pub const LINUX_MALLOC: usize = 2;
 
 // TODO set proper values
